fix(dcpp): fix some compile errors that happen with -DNO_ASM

+++ src/dcpp/cpp.c
@@ -126,7 +126,7 @@ int32_t xbytes;
                 }
             }
 #else
-            len = HandleTriGraphs(base) - (int)base;
+            len = (int)HandleTriGraphs(base) - (int)base;
 #endif
         }
     } else {             /*  else macro  */
